Cache NeoAi::Client at class level and extract NeoAiLessonSerializer concern
- Use class-level memoization in BaseController so the JWT token cache
  survives across requests instead of being discarded per request
- Extract duplicate serialize_lesson, serialize_scene, and
  derive_lesson_status into a shared NeoAiLessonSerializer concern
  included in BaseController
